WL: break out subsurface.py
